% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_obs_in_window.R
\name{check_obs_in_window}
\alias{check_obs_in_window}
\title{Checks if there a observations in a given window}
\usage{
check_obs_in_window(i, prcp, delta_window = 30)
}
\arguments{
\item{i}{index of the observation}

\item{prcp}{a vector of daily precpitation observations (these are assumed sequential)}

\item{delta_window}{window to check, i is in the centre of this window (default = 30)}
}
\value{
Returns a boolean if there a observations in the window
}
\description{
Checks if there a observations in a given window
}
\examples{
prcp = 1:365
prcp[100:120] = NA
check_obs_in_window(i = 100, prcp)
check_obs_in_window(i = seq(10, 360, 50), prcp, delta_window = 10)
check_obs_in_window(i = seq(10, 360, 50), prcp)
}
