% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_king_test.R
\name{extremes_untagged_test}
\alias{extremes_untagged_test}
\title{Checks that the frequency of days rainfall extremes are observed is equal
throughout the days of the week, (King 2013)
Test runs on yearly observations, so the corresponding block index is the year.}
\usage{
extremes_untagged_test(dates, alpha = 0.05, min_obs = 40)
}
\arguments{
\item{dates}{vector of dates}

\item{alpha}{alpha level of the two-sided hypothesis test (default = 0.05)}

\item{min_obs}{minimum number of observations needed to run the test (default = 40)}
}
\value{
Returns the a data frame with columns of wday for weekday, freq for frequency observation,
lower for whether the freq is below the low confidence interval,
upper for when the freq of the next day is above the confidence interval,
bool for when both the lower and upper indicate that there are untagged accumulations
}
\description{
Checks that the frequency of days rainfall extremes are observed is equal
throughout the days of the week, (King 2013)
Test runs on yearly observations, so the corresponding block index is the year.
}
\examples{

dates = seq(as.Date("1910-01-01"), as.Date("1910-12-31"), by ="days")
extremes_untagged_test(dates)
monday_date = dates[wday(dates) == 2][1]
sunday_ind = which(wday(dates) == 1)
samp = sample(sunday_ind, 0.3*length(sunday_ind))
dates[samp] = monday_date
extremes_untagged_test(dates)

}
