% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_block_index.R
\name{get_block_index}
\alias{get_block_index}
\title{Get index of the block associated with a given date}
\usage{
get_block_index(dates, start_month, end_month)
}
\arguments{
\item{dates}{vector of dates}

\item{start_month}{beginning of the block}

\item{end_month}{end of the block}
}
\value{
Returns the block index for the given date.
The block index is a year, and is the year of the \code{start_month}.
}
\description{
Get index of the block associated with a given date
}
\examples{
get_block_index(as.Date("1910-01-01"), 1, 12)
get_block_index(as.Date("1910-01-01"), 12, 2)

}
