% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_block_length.R
\name{get_block_length}
\alias{get_block_length}
\title{Get the length of the block}
\usage{
get_block_length(start_month, end_month)
}
\arguments{
\item{start_month}{beginning of the block.}

\item{end_month}{end of the block.}
}
\value{
The number of days between the \code{start_month} and the \code{end_month}. Function ignores February 29.
}
\description{
Get the length of the block
}
\examples{
get_block_length(1, 12)
get_block_length(12, 2)
}
