% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_block_max.R
\name{get_block_max}
\alias{get_block_max}
\title{Get the highest observations in the block}
\usage{
get_block_max(stn_prcp, block_val, rmax = 1)
}
\arguments{
\item{stn_prcp}{columns of block, date, prcp}

\item{block_val}{block to get the maxima from}

\item{rmax}{number of highest observations to get per block (default = 1)}
}
\value{
Returns the highest observations in the block matching \code{block_val}
The returned data frame has columns of block, data and prcp.
If insufficient data, then \code{NULL} is returned
}
\description{
Get the highest observations in the block
}
\examples{
prcp = c(1:365, 2*(1:365))
dates = seq(as.Date("1910-01-01"), as.Date("1911-12-31"), by = "days")
block = get_block_index(dates, 1, 12)
stn_prcp = data.frame(block, date = dates, prcp)
get_block_max(stn_prcp, 1910)
get_block_max(stn_prcp, 1910, rmax = 4)
block = get_block_index(dates, 3, 5)
stn_prcp = data.frame(block, date = dates, prcp)
get_block_max(stn_prcp, 1910)
get_block_max(stn_prcp, 1910, rmax = 4)
}
