% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stn_max.R
\name{get_stn_max}
\alias{get_stn_max}
\title{Get the highest observations in the each block}
\usage{
get_stn_max(stn_prcp, stn_id, start_month, end_month, min_perc = 0.95,
  rmax = 1, dates = NULL, prcp = NULL)
}
\arguments{
\item{stn_prcp}{columns of block, date, prcp}

\item{stn_id}{(optional) covienient for indexing}

\item{start_month}{beginning of the block.}

\item{end_month}{end of the block.}

\item{min_perc}{minimum percentage of missing data per block (default = 0.95)}

\item{rmax}{number of highest observations to get per block (default = 1)}

\item{dates}{(optional instead of stn_prcp)
vector of dates}

\item{prcp}{(optional instead of stn_prcp)
vector of precipitation}
}
\value{
Returns the a data frame with columns id, block, date, prcp, num_obs, max
the prcp is the raw maxima, date is the date of the maxima,
num_obs is the number of observations per block, and
max is set to \code{NA} when there are not enough num_obs
observations per block based on the \code{min_perc}
}
\description{
Get the highest observations in the each block
}
\examples{
prcp = c(1:365, 2*(1:365))
dates = seq(as.Date("1910-01-01"), as.Date("1911-12-31"), by = "days")
block = get_block_index(dates, 1, 12)
stn_prcp = data.frame(block, date = dates, prcp)
get_stn_max(stn_prcp, "STN_ID", 1, 12)
get_stn_max(stn_prcp, "STN_ID", 1, 12, rmax = 2)
block = get_block_index(dates, 3, 5)
stn_prcp = data.frame(block, date = dates, prcp)
get_stn_max(stn_prcp, "STN_ID", 3, 5)
get_stn_max(stn_prcp, "STN_ID", 3, 5, rmax = 2)
block = get_block_index(dates, 12, 2)
stn_prcp = data.frame(block, date = dates, prcp)
get_stn_max(stn_prcp, "STN_ID", 12, 2)
get_stn_max(stn_prcp, "STN_ID", 12, 2, min_perc = 0.5)

}
