% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WORKING_plot_rainfall_spatially.R
\name{plot_dates_spatially}
\alias{plot_dates_spatially}
\title{Produces a spatial plot for a given set of dates and observations}
\usage{
plot_dates_spatially(date_range, prcp_var, mdpr_var, dapr_var,
  stn_meta_data)
}
\arguments{
\item{date_range}{dates for plotting}

\item{prcp_var}{standard rnoaa format for precipiation, PRCP}

\item{mdpr_var}{standard rnoaa format for accumulated precipitation totals, MDPR}

\item{dapr_var}{standard rnoaa format for days accumulated precipitation, DAPR}

\item{stn_meta_data}{the associated meta data for the stations,
must have columns of longitude and latitude}
}
\value{
Returns the plot
}
\description{
Produces a spatial plot for a given set of dates and observations
}
\examples{

}
