% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recon_row.R
\name{recon_row}
\alias{recon_row}
\title{Find first observations that is not NA in the row}
\usage{
recon_row(df_row)
}
\arguments{
\item{df_row}{a row of precipitation observations,
first column is date, second columns is the station to be reconstructed,
 and the other rownames are neighbouring station ids.
 The values in all columns except the first are prcp.}
}
\value{
Returns a data frame with columns of
date, id, recon_prcp and recon_flag
}
\description{
Find first observations that is not NA in the row
}
\examples{
date = seq(as.Date("1910-01-01"), as.Date("1910-01-12"), by = "days")
id1 = rexp(length(date), rate = 1)
id2 = id1 + rnorm(length(date), 0, 1)
id3 = id1 + rnorm(length(date), 0.5, 2)
prcp_df = data.frame(date, id1, id2, id3)
prcp_df$id1[1:10] = NA
prcp_df$id2[1:5] = NA
prcp_df$id3[1] = NA
recon_row(prcp_df[1,])
recon_row(prcp_df[2,])
recon_row(prcp_df[6,])
recon_row(prcp_df[11,])
recon_df = NULL
for(i in 1:length(date)){
  recon_df = rbind(recon_df, recon_row(prcp_df[i, ]))
}
recon_df
### APPLY / SAPPLY ARE INHERITING NAMES WEIRDLY
### apply(prcp_df[1:3, ], 1, recon_row)
### This creates an error
}
