% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recon_stn.R
\name{recon_stn}
\alias{recon_stn}
\title{Reconstructs missing observations at a station by finding the
first observation in each row that is not NA}
\usage{
recon_stn(prcp_df, delta_window)
}
\arguments{
\item{prcp_df}{a df of precipitation observations,
first column is date, second columns is the station to be reconstructed,
 and the other rownames are neighbouring station ids.
 The values the first column are dates and
 the values are precipitation in all columns}
}
\value{
Returns a data frame with columns of
date, id, recon_prcp and recon_flag
}
\description{
Reconstructs missing observations at a station by finding the
first observation in each row that is not NA
}
\examples{
date = seq(as.Date("1910-01-01"), as.Date("1910-01-31"), by = "days")
id1 = rexp(length(date), rate = 1)
id2 = id1 + rnorm(length(date), 0, 1)
id3 = id1 + rnorm(length(date), 0.5, 2)
prcp_df = data.frame(date, id1, id2, id3)
View(recon_stn(prcp_df))
prcp_df$id1[1:10] = NA
prcp_df$id2[1:5] = NA
prcp_df$id3[1] = NA
View(recon_stn(prcp_df))
View(recon_stn(prcp_df, delta_window = 3))
}
