% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_max_using_neighbours.R
\name{update_max_using_neighbours}
\alias{update_max_using_neighbours}
\title{Checks if the date of the missing maximum is similar to neighbours.
If it is, them it keeps the maxima despite the missing data.}
\usage{
update_max_using_neighbours(meta_data, stn_id, max_var,
  search_radius = 20)
}
\arguments{
\item{meta_data}{meta data of the stations}

\item{stn_id}{index of the station}

\item{max_var}{data frame with columns id, date, prcp, block, num_obs, max
output from \code{get_stn_max()}. The prcp col is the raw maxima,
date is the date of the maxima, num_obs is the number of observations per block,
and max is set to NA when there are not enough num_obs observations per block
based on the \code{min_perc} set in \code{get_stn_max()}.}

\item{search_radius}{(default = 20) radius of neighbours to compare dates}
}
\value{
Returns the updated station max object that keeps the missing maxima
if the date is common with the neighbours.
}
\description{
Checks if the date of the missing maximum is similar to neighbours.
If it is, them it keeps the maxima despite the missing data.
}
\examples{
stn_id = "ASN00040383"
nbr_id = "ASN00040225"
meta_data = readRDS("Data/AS_meta_data.RDS")
stn_dates = c(as.Date("1966-01-21"), as.Date("1967-03-05"), as.Date("1968-02-02"))
stn_block = get_block_index(stn_dates, 1, 12)
stn_prcp = c(100, 200, 300)
stn_max = c(NA, 200, NA)
stn_num_obs = c(100, 365, 100)
max_var_stn = data.frame(id = rep(stn_id, 3),
date = stn_dates, prcp = stn_prcp, block = stn_block, num_obs = stn_num_obs,
max = stn_max)
nbr_dates = c(as.Date("1966-01-21"), as.Date("1967-03-05"), as.Date("1968-02-02"))
nbr_block = get_block_index(nbr_dates, 1, 12)
nbr_prcp = c(101, 201, NA)
nbr_max = nbr_prcp
nbr_num_obs = c(365, 365, 0)
max_var_nbr = data.frame(id = rep(nbr_id, 3),
date = nbr_dates, prcp = nbr_prcp, block = nbr_block, num_obs = nbr_num_obs,
max = nbr_max)
max_var = rbind(max_var_stn, max_var_nbr)
update_max_using_neighbours(meta_data, stn_id, max_var)

}
