% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_correlation.R
\name{utils_correlation}
\alias{utils_correlation}
\title{Get the correlation between two stations}
\usage{
utils_correlation(prcp_df, stn_pair, type = "pearson", min_overlap = 365 *
  3)
}
\arguments{
\item{prcp_df}{has columns of stn ids, but no date just precipitation data}

\item{stn_pair}{is a pair of stn ids - caution input must be characters!!!!}

\item{type}{correlation type, default is "pearson" otherwise, "spearman"}

\item{min_overlap}{is the number of observations needed before calculating corr(x,y)}
}
\value{
Returns the correlation between the observation at the two stations
}
\description{
Get the correlation between two stations
}
\examples{
id1 = rexp(365*10, rate = 1)
id2 = id1 + rnorm(365*10, 0, 1)
prcp_df = data.frame(id1, id2)
stn_pair = c("id1", "id2")
utils_correlation(prcp_df, stn_pair)
utils_correlation(prcp_df, stn_pair, type = "spearman")
utils_correlation(prcp_df, stn_pair, type = "spearman", min_overlap = 100)
}
