% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_get_meta_data.R
\name{wrapper_get_meta_data}
\alias{wrapper_get_meta_data}
\title{Get station meta data from GHCN-Daily (defaults to Australia)}
\usage{
wrapper_get_meta_data(country_id = "AS", network_id = "N")
}
\arguments{
\item{country_id}{(default is AS for Australia) see RNOAA package for other ids}

\item{network_id}{(default is N for Australia Bureau of Meteorology) see RNOAA package for other ids}
}
\value{
Returns the Australia station meta data for all possible stations
}
\description{
Get station meta data from GHCN-Daily (defaults to Australia)
}
