% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_save_prcp_data.R
\name{wrapper_save_prcp_data}
\alias{wrapper_save_prcp_data}
\title{Gets station data from GHCN-Daily for a given set of stations IDs}
\usage{
wrapper_save_prcp_data(stn_ids, data_dir, file_str = "", ...)
}
\arguments{
\item{stn_ids}{a set of station ids}

\item{data_dir}{the directory to save the data}

\item{file_str}{a reference string for the file (default = "")}
}
\value{
Returns NULL, but saves the data for all station ids into
separate files for each of the element types. The outpute files take the form
paste(data_dir, element_type, file_str, ".rds", sep ="")
}
\description{
Can also pass all inputs from rnoaa:meteo_pull_monitors()
}
