% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_viney.R
\name{wrapper_viney}
\alias{wrapper_viney}
\title{Wrapper for test for Sunday Monday Untagged Accumulations}
\usage{
wrapper_viney(stn_ids, prcp_var, dapr_var)
}
\arguments{
\item{stn_ids}{vector station ids to run the test}

\item{prcp_var}{standard rnoaa format for precipiation, PRCP,
see rnoaa::meteo_pull_monitors()}

\item{dapr_var}{standard rnoaa format for days accumulated precipitaiotn, DAPR,
see rnoaa::meteo_pull_monitors()}
}
\value{
Returns the a data frame with columns block, which is the year, and p_value.
}
\description{
Checks for the presence of untagged Sunday Monday accumulations as in
Viney and Bates (2004).
}
\details{
Test currently runs on yearly observations only. The block index returned
therefore corresponds to the year.
}
