% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_fevd_with_tryCatch.R
\name{fit_fevd_with_tryCatch}
\alias{fit_fevd_with_tryCatch}
\title{Fits a GEV distribution and returns the parameters
This is a wrapper to handle any errors}
\usage{
fit_fevd_with_tryCatch(x)
}
\arguments{
\item{x}{a vector of block maximum observations}
}
\description{
Fits a GEV distribution and returns the parameters
This is a wrapper to handle any errors
}
\examples{
x1 = revd(100,1,1,1)
par1 = fit_fevd_with_tryCatch(x1)
x2 = "SOMETHING SILLY"
par2 = fit_fevd_with_tryCatch(x2)
}
